fix: swal handler

+++ src/store/index.js
@@ -13,7 +13,8 @@ export default new Vuex.Store({
     isLoggedIn: false,
     categoryName: '',
     carts: [],
-    histories: []
+    histories: [],
+    userLoggedIn: ''
   },
   mutations: {
     SET_BANNERS (state, payload) {
@@ -36,6 +37,9 @@ export default new Vuex.Store({
     },
     SET_HISTORIES (state, payload) {
       state.histories = payload
+    },
+    SET_USER_LOGGED_IN (state, payload) {
+      state.userLoggedIn = payload
     }
   },
   actions: {
@@ -86,7 +90,9 @@ export default new Vuex.Store({
       })
         .then(({ data }) => {
           context.commit('SET_ISLOGGEDIN', true)
+          context.commit('SET_USER_LOGGED_IN', data.email)
           localStorage.setItem('access_token', data.access_token)
+          localStorage.setItem('email', data.email)
           router.push('/')
           Vue.swal({
             title: 'Success login!',
@@ -122,10 +128,32 @@ export default new Vuex.Store({
         })
     },
     handleLogout (context, payload) {
-      localStorage.clear()
-      context.commit('SET_ISLOGGEDIN', payload)
-      this.state.categoryName = ''
-      router.push('/')
+      Vue.swal({
+        title: 'Are you sure?',
+        text: 'Are you sure want to logout?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yep!'
+      })
+        .then(res => {
+          if (res.isConfirmed) {
+            localStorage.clear()
+            context.commit('SET_ISLOGGEDIN', payload)
+            this.state.categoryName = ''
+            router.push('/')
+            Vue.swal({
+              title: 'Logout success!',
+              icon: 'success',
+              toast: true,
+              position: 'top-end',
+              timer: 2000,
+              timerProgressBar: true,
+              showConfirmButton: false
+            })
+          }
+        })
     },
     handleAddProductToCart (context, payload) {
       axios({
@@ -143,12 +171,14 @@ export default new Vuex.Store({
           context.dispatch('fetchCarts')
         })
         .catch(err => {
+          console.log('ALLLO')
           console.log(err.response)
           Vue.swal({
             title: 'Oopsie!',
-            text: err.response.data.message,
+            text: 'Please login first',
             icon: 'error'
           })
+          router.push('/login')
         })
     },
     fetchCarts (context, payload) {
@@ -206,6 +236,7 @@ export default new Vuex.Store({
             timerProgressBar: true,
             showConfirmButton: false
           })
+          router.push('/history')
         })
         .catch(err => {
           console.log(err.response)
